eunice-data-engine/requirements.txt
python-gitlab>=4.0.0
pyyaml>=6.0
requests>=2.31.0
